+++ CHANGELOG.md
@@ -9,10 +9,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support `AddSlideOut` extension.
 - Support `SetDialogLauncher` and `GetDialogLauncher` extensions.
 - Support `CreateToggleButton` extension.
+- Support `CreateTextBox` extension.
+- Support `CreateLabel` extension.
 ### UI
 - Update `RibbonExtension` to support `AddSlideOut`.
 - Update `RibbonExtension` to support `SetDialogLauncher` and `GetDialogLauncher`.
 - Update `RibbonExtension` to support `CreateToggleButton`.
+- Update `RibbonExtension` to support `CreateTextBox`.
+- Update `RibbonExtension` to support `CreateLabel`.
 
 ## [0.3.0] / 2025-07-15
 ### Features